cielo-rio-grande/backend/requirements.txt
fastapi
uvicorn[standard]
requests
opencv-python
python-multipart
tensorflow
pillow
numpy
python-dotenv
typing_extensions
apscheduler
pytz